requirements.txt
bcrypt==4.1.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
idna==3.6
paramiko==3.4.0
pycparser==2.21
pyfgt==0.5.6
PyNaCl==1.5.0
PyYAML==6.0.1
requests==2.31.0
str2bool==1.1
urllib3==2.1.0
